# src/main/resources/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Book Store",
    "description": "書店商品API",
    "termsOfService": "https://github.com/ntub46010/SpringBootTutorial/tree/Ch8",
    "version": "Ver. 1.0.2"
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Product",
      "description": "Management product"
    }
  ],
  "paths": {
    "/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "#putApi_S2",
        "operationId": "replaceProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "#deleteApi_S3",
        "description": "Delete a product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "getProducts",
        "parameters": [
          {
            "name": "param",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProductQueryParameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "#postApi_S1",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProductQueryParameter": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string"
          },
          "priceFrom": {
            "type": "integer",
            "format": "int32"
          },
          "priceTo": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "string"
          },
          "sortRule": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}